All functionality implemented
Needs more testing of multiple users and the to-do list, but all functionality *should* be implemented.

The method to use here is
Sync.sync(DateTime.now());
whenever you write to globals, and
Sync.sync(); to update either the local app or the database to the most recent changes. The latter should be called on startup, and maybe every minute or every few minutes.

Current problem logically is that there is the potential for a user to try to sync changes before they get the most up-to-date database, leading to data loss. Could maybe be solved with a listener on the database and syncing on any change.